refactor lodash
