+++ src/app/features/feature-movie/components/movie-form/movie-form.component.ts
@@ -1,4 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ChangeDetectorRef, Output, EventEmitter } from '@angular/core';
+import { v4 as uuid } from 'uuid';
+import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { Movie } from '../../models/movie';
 
 @Component({
   selector: 'app-movie-form',
@@ -7,9 +10,55 @@ import { Component, OnInit } from '@angular/core';
 })
 export class MovieFormComponent implements OnInit {
 
-  constructor() { }
+  @Output() onNewMovie = new EventEmitter<Movie>();
+  
+  public movieForm: FormGroup;
 
-  ngOnInit(): void {
+  constructor(
+    private _changeDetector: ChangeDetectorRef,
+  ) {
+    this.movieForm = new FormGroup({
+      'title': new FormControl('', [Validators.required, Validators.minLength(4), Validators.maxLength(30)]),
+      'release': new FormControl('', Validators.required),
+      'image': new FormControl(null, Validators.required),
+      'description': new FormControl('', Validators.compose([Validators.required, Validators.minLength(15), Validators.maxLength(500)]))
+    });
+  }
+
+  ngOnInit(): void { }
+
+  get title() { return this.movieForm.get('title'); }
+  get release() { return this.movieForm.get('release'); }
+  get image() { return this.movieForm.get('image'); }
+  get description() { return this.movieForm.get('description'); }
+
+  onFileChange(event) {
+    let reader = new FileReader();
+
+    if (event.target.files && event.target.files.length) {
+      const [file] = event.target.files;
+      reader.readAsDataURL(file);
+
+      reader.onload = () => {
+        this.movieForm.patchValue({
+          imageMovie: reader.result
+        });
+
+        this._changeDetector.markForCheck();
+      };
+    }
+  }
+
+  onSubmitMovie() {
+    const form = this.movieForm && this.movieForm.value;
+    const payload = {
+      id: uuid(),
+      title: form.title,
+      release: form.release,
+      image: form.image,
+      description: form.description
+    };
+    this.onNewMovie.emit(payload)
   }
 
 }
